Refactored psort front-end to tool #160
