--- tests/frontend/psort.py
@@ -1,17 +1,0 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
-"""Tests for the psort front-end."""
-
-import unittest
-
-from tests import test_lib as shared_test_lib
-
-
-class PsortFrontendTest(shared_test_lib.BaseTestCase):
-  """Tests for the psort front-end."""
-
-  # TODO: implement.
-
-
-if __name__ == '__main__':
-  unittest.main()
